Added ability to import mqo from mqoz file
Metasequoia 4.7 saves to .mqoz format which is a zip file containg an .mqo file and a thumbnail.jpeg and maybe a .mqx xml file. Modified the importer to load the .mqo in the zip. Registered undo for importer.  Added some error checking to importer if user  turns off file filtering in file browser and selects non .mqo or .mqoz.